adding context and ability to give less than 1 refresh intervals

+++ migrations/V1__create_alerts_table.sql
@@ -5,8 +5,8 @@ CREATE TABLE uphold_alerts.alerts (
       pair VARCHAR(20) NOT NULL,
       price_change NUMERIC(30, 20) NOT NULL,
       perc_change NUMERIC(30, 20) NOT NULL,
-      final_price numeric(30, 20) NOT NULL,
-      config_refresh INT NOT NULL,
+      final_price NUMERIC(30, 20) NOT NULL,
+      config_refresh NUMERIC(10, 5) NOT NULL,
       config_perc_oscillation NUMERIC(20, 10) NOT NULL,
       timestamp TIMESTAMP NOT NULL
 );
